combine default clients
