+++ se_gym/client.py
@@ -6,6 +6,7 @@ import typing
 
 __all__ = ["set_client"]
 
+
 class _Client:
     _instance = None
     _initialized = False
@@ -45,5 +46,24 @@ class _Client:
         else:
             return getattr(resp, field_name)
 
+
 def set_client(client: openai.Client):
     _Client(client=client)
+
+
+def LMU_get_openai_client():
+    import openai
+    import httpx
+    import os
+    import dotenv
+    from . import utils
+
+    dotenv.load_dotenv()
+    USERNAME = os.getenv("API_USERNAME")
+    PASSWORD = os.getenv("API_PASSWORD")
+    assert USERNAME is not None, "API_USERNAME is not set in .env or file not found."
+    assert PASSWORD is not None, "API_PASSWORD is not set in .env or file not found."
+    openai.OpenAI.custom_auth = httpx.BasicAuth(USERNAME, PASSWORD)
+    client = openai.OpenAI(base_url="https://ollama.mobile.ifi.lmu.de/v1/", api_key="none")
+    utils.check_client(client)
+    return client
